修改页数排序
# message auto-generated for no-merge-commit merge:
merge paixu into master

增加communication后端与前端交互接口排序功能

Created-by: insightDeveloper

Match-id-1e8bfabc873f30e793f4acc576f7aae05c185c36

+++ server/src/modules/timeline/protocol/TimelineProtocolRequest.h
@@ -124,8 +124,8 @@ struct SearchSliceRequest : public Request {
 
 struct CommunicationDetailParams {
     std::string rankId;
-    double currentPage = 0;
-    double pageSize = 0;
+    int32_t currentPage = 0;
+    int32_t pageSize = 0;
     std::string orderBy;
     std::string order;
 };
